requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cryptography==44.0.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
duckduckgo_search==8.0.1
dydantic==0.0.8
executing==2.2.0
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-google-genai==2.1.4
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-api==0.2.11
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.8
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.9
langgraph-sdk==0.1.66
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.38.0
numpy==2.2.5
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
structlog==25.3.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
trustcall==0.0.39
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cryptography==44.0.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
duckduckgo_search==8.0.1
dydantic==0.0.8
executing==2.2.0
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-google-genai==2.1.4
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-api==0.2.11
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.8
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.9
langgraph-sdk==0.1.66
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.38.0
numpy==2.2.5
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
structlog==25.3.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
trustcall==0.0.39
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
